Refactor: Move all unit test to test directory.
