+++ src/quantum_info/observable.rs
@@ -356,48 +356,3 @@ impl Drop for Observable {
         unsafe { qiskit_sys::qk_obs_free(self.observable) };
     }
 }
-
-#[cfg(test)]
-mod observable_tests {
-    use crate::quantum_info::observable::BitTerm;
-
-    use super::{Complex64, Observable};
-    #[test]
-    fn test_new_observable() {
-        let num_qubits = 100;
-        let coeffs = [
-            Complex64 { re: 1.0, im: -1.0 },
-            Complex64 { re: 1.0, im: -1.0 },
-        ];
-        let bits = [BitTerm::Zero, BitTerm::One, BitTerm::Plus, BitTerm::Minus];
-        let indices = [0, 1, 98, 99];
-        let boundaries = [0, 2, 4];
-
-        let obs = Observable::new(num_qubits, &coeffs, &bits, &indices, &boundaries);
-        assert_eq!(obs.num_terms(), 2);
-        assert_eq!(obs.coeffs().len(), coeffs.len());
-        for (i, coef) in obs.coeffs().enumerate() {
-            assert_eq!(coef.re, coeffs[i].re);
-            assert_eq!(coef.im, coeffs[i].im);
-        }
-        assert_eq!(obs.indices().len(), indices.len());
-        for (i, idx) in obs.indices().enumerate() {
-            assert_eq!(*idx, indices[i]);
-        }
-        assert_eq!(obs.boundaries().len(), boundaries.len());
-        for (i, idx) in obs.boundaries().enumerate() {
-            assert_eq!(*idx, boundaries[i]);
-        }
-
-        assert_eq!(obs.bit_terms().len(), bits.len());
-        for (i, idx) in obs.bit_terms().as_slice().iter().enumerate() {
-            assert_eq!(*idx, bits[i]);
-        }
-
-        assert_eq!(obs.coeffs().len() + 1, obs.boundaries().len());
-        assert_eq!(obs.bit_terms().len(), obs.indices().len());
-
-        let obs_b = Observable::new(num_qubits, &coeffs, &bits, &indices, &boundaries);
-        assert!(obs.equal(&obs_b));
-    }
-}
