+++ tests/circuit/test_registers.rs
